Ver: 0.4.1 alpha Feat: Implement dialog to confirm passenger on bus
This commit introduces a dialog box that prompts the user to confirm whether they are currently on the bus. This dialog is displayed upon entering the bus route map screen.
The dialog provides options for the user to indicate if they are on the bus or not. This information is then used to control whether the user's location is continuously tracked on the map.

The  widget has been added and integrated into the  to provide this functionality.